C8Pms_Tai_Hw5.0_Fw1.0.0.1750
1、BLE数据改由小端模式通信
2、设备类型总是发送

+++ Src/source/JT808/JtTlv0900.h
@@ -121,7 +121,7 @@ extern "C"{
 		uint16_t	Minor ;
 		int8_t		RSSI ;
 		uint8_t 	soc ;		// 电池百分比
-		uint8_t		Voltage ;	// 电池电压
+		uint16_t	Voltage ;	// 电池电压
 	}Beacon;
 
 	typedef struct
